# qiansi-server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "纸喵软件系列之服务端",
        "title": "纸喵 qiansi API",
        "termsOfService": "http://zhimiao.org",
        "contact": {
            "name": "API Support",
            "url": "http://tools.zhimiao.org",
            "email": "mail@xiaoliu.org"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:1315",
    "basePath": "/",
    "paths": {
...
        "/admin/UserResetPwd": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "修改密码",
                "parameters": [
                    {
                        "description": "入参集合",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/req.UserResetPwdParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 1,\"msg\": \"修改成功\", \"data\": null}",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/resp.ApiResult"
                        }
                    }
                }
            }
        },
        "/admin/UserSigin": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "description": "入参集合",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/req.UserSiginParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 1,\"msg\": \"登录成功\", \"data\": {\"CreateTime\": \"2019-02-27T16:11:27+08:00\",\"InviterUid\": 0,\"Password\": \"\",\"Phone\": \"15061370322\",\"Status\": 1,\"Uid\": 2, \"UpdateTime\": \"2019-02-27T16:19:54+08:00\", \"Token\":\"sdfsdafsd..\"}}",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/resp.ApiResult"
                        }
                    }
                }
            }
        },
        "/admin/UserSiginUp": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "注册账号",
                "parameters": [
                    {
                        "description": "入参集合",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/req.UserSiginUpParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 1,\"msg\": \"注册成功\",\"data\": {\"CreateTime\": \"2019-02-27T16:11:27+08:00\",\"InviterUid\": 0,\"Password\": \"\",\"Phone\": \"15061370322\",\"Status\": 1,\"Uid\": 2, \"UpdateTime\": \"2019-02-27T16:19:54+08:00\", \"Token\":\"sdfsdafsd..\"}}",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/resp.UserInfoVO"
                        }
                    }
                }
            }
        },
...
        "/admin/VerifyBySMS": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取短信验证码",
                "parameters": [
                    {
                        "description": "入参集合",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/req.VerifyBySMSParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":1,\"msg\":\"发送成功\",\"data\":null}",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/resp.ApiResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "req.UserResetPwdParam": {
            "type": "object",
            "properties": {
                "newPassword": {
                    "type": "string"
                },
                "oldPassword": {
                    "type": "string"
                }
            }
        },
        "req.UserSiginParam": {
            "type": "object",
            "properties": {
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "phone": {
                    "description": "手机号",
                    "type": "string"
                }
            }
        },
        "req.UserSiginUpParam": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "短信验证码",
                    "type": "string"
                },
                "inviterUid": {
                    "description": "邀请人",
                    "type": "integer"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "phone": {
                    "description": "手机号",
                    "type": "string"
                }
            }
        },
        "req.VerifyBySMSParam": {
            "type": "object",
            "properties": {
                "imgCode": {
                    "description": "验证码code",
                    "type": "string"
                },
                "imgIdKey": {
                    "description": "图片验证码id",
                    "type": "string"
                },
                "phone": {
                    "description": "手机号",
                    "type": "string"
                }
            }
        },
        "resp.ApiResult": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "resp.UserInfoVO": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        }
    }
}